refactor(zaino-state): extract the txout-set accumulator into its own module
Move the finalised txout-set accumulator implementation out of
transparent_address_history.rs (a different capability, gated by a different
feature) into a dedicated tx_out_set_accumulator.rs, declared as a single
`#[cfg(feature = "gettxoutsetinfo")] pub(crate) mod`. With the whole module
gated, no item inside needs its own feature cfg.

Moved: the `AccumulatorDirection` enum and the free delta helpers
(apply_tx_out_set_entries_delta, apply_in_block_transitions, apply_entry_deltas,
index_created_outputs, index_spent_outpoints); the private accumulator methods
(resolve_spent_outpoints_for_set_info, apply_prior_block_transitions,
build_entry_data); the read / write-path-delta / rebuild methods — all now in a
single `impl DbV1` — and the accumulator helper unit tests (`#[cfg(test)] mod`).

Stays behind: the `impl TransparentHistExt for DbV1` forwarder for
get_tx_out_set_info_accumulator. A trait impl is coherence-unique and lives with
the address-history methods, and the trait method must exist even with the
feature off (to return FeatureUnavailable); the thin shim delegates to the
inherent method, now `pub(super)`. `find_txout_in_stored_transparent_tx_list` is
shared with the write path and stays, bumped to `pub(super)` for a moved test.

Behaviour-preserving: transparent_address_history.rs 2515 -> ~1030 lines. Lib
and tests build clean and pass clippy in both feature states.
